feat: add monthly income data to email object
The changes in this commit add the `monthlyIncomeData` property to the email object in the `mails` array. This property contains an array of monthly income data, including the income, change, and limit for each month. The `generateMonthlyIncomeData` function is used to generate this data based on a base income value. This addition allows for tracking and displaying monthly income data in the email application.
chore: update dependencies

- Update "turbo" dependency from version 1.11.3 to 1.12.4 in package.json
- Update "turbo" dependency from version 1.11.3 to 1.12.4 in client code
- Update "@trpc/server" dependency from version 11.0.0-next-beta.241 to 11.0.0-next-beta.242 in yarn.lock
- Update "turbo-darwin-64" dependency from version 1.11.3 to 1.12.4 in yarn.lock
- Update "turbo-darwin-arm64" dependency from version 1.11.3 to 1.12.4 in yarn.lock
- Update "turbo-linux-64" dependency from version 1.11.3 to 1.12.4 in yarn.lock
- Update "turbo-linux-arm64" dependency from version 1.11