feat: add EmptyScreen component to display welcome message and example messages for new users
The EmptyScreen component is added to provide a welcoming experience for new users. It displays a Card with a title and description, along with a list of example messages that users can click on to get started with using the app. This component is added to the main screen of the application to guide users in navigating their finances with ease.